Move all EffectHandlers to presentation/handler package
- Create handler/ subdirectory in each feature's presentation package
- Move all *EffectHandler.kt files to handler/ subdirectory
- Update package declarations and imports
- Update factory and test imports
- Fix FlowHelpersTest to follow `when ... then ...` naming convention

Improves organization and follows single responsibility principle for package structure.
