+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/app/di/AppFeatureFactory.kt
@@ -4,8 +4,8 @@ import kotlinx.datetime.LocalDate
 import space.be1ski.vibits.shared.app.domain.model.AppState
 import space.be1ski.vibits.shared.app.presentation.AppAction
 import space.be1ski.vibits.shared.app.presentation.AppEffect
-import space.be1ski.vibits.shared.app.presentation.AppEffectHandler
 import space.be1ski.vibits.shared.app.presentation.appReducer
+import space.be1ski.vibits.shared.app.presentation.handler.AppEffectHandler
 import space.be1ski.vibits.shared.core.elm.Feature
 import space.be1ski.vibits.shared.core.elm.FeatureImpl
 import space.be1ski.vibits.shared.feature.mode.domain.model.AppMode
