+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/feature/habits/presentation/handler/HabitsActivityEffectHandler.kt
@@ -1,4 +1,4 @@
-package space.be1ski.vibits.shared.feature.habits.presentation
+package space.be1ski.vibits.shared.feature.habits.presentation.handler
 
 import kotlinx.coroutines.flow.Flow
 import space.be1ski.vibits.shared.core.elm.EffectHandler
@@ -6,6 +6,8 @@ import space.be1ski.vibits.shared.core.elm.actions
 import space.be1ski.vibits.shared.core.logging.Log
 import space.be1ski.vibits.shared.feature.habits.domain.usecase.CalculateActivityDataUseCase
 import space.be1ski.vibits.shared.feature.habits.domain.usecase.PrewarmActivityDataUseCase
+import space.be1ski.vibits.shared.feature.habits.presentation.HabitsAction
+import space.be1ski.vibits.shared.feature.habits.presentation.HabitsEffect
 
 private const val TAG = "HabitsActivityEffect"
