+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/feature/habits/presentation/handler/HabitsEffectHandler.kt
@@ -1,7 +1,9 @@
-package space.be1ski.vibits.shared.feature.habits.presentation
+package space.be1ski.vibits.shared.feature.habits.presentation.handler
 
 import kotlinx.coroutines.flow.Flow
 import space.be1ski.vibits.shared.core.elm.EffectHandler
+import space.be1ski.vibits.shared.feature.habits.presentation.HabitsAction
+import space.be1ski.vibits.shared.feature.habits.presentation.HabitsEffect
 
 class HabitsEffectHandler(
   private val memoHandler: HabitsMemoEffectHandler,
