+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/feature/memos/presentation/handler/MemosCredentialsEffectHandler.kt
@@ -1,4 +1,4 @@
-package space.be1ski.vibits.shared.feature.memos.presentation
+package space.be1ski.vibits.shared.feature.memos.presentation.handler
 
 import kotlinx.coroutines.flow.Flow
 import space.be1ski.vibits.shared.core.elm.EffectHandler
@@ -6,6 +6,8 @@ import space.be1ski.vibits.shared.core.elm.actions
 import space.be1ski.vibits.shared.feature.auth.domain.model.Credentials
 import space.be1ski.vibits.shared.feature.auth.domain.usecase.LoadCredentialsUseCase
 import space.be1ski.vibits.shared.feature.auth.domain.usecase.SaveCredentialsUseCase
+import space.be1ski.vibits.shared.feature.memos.presentation.MemosAction
+import space.be1ski.vibits.shared.feature.memos.presentation.MemosEffect
 
 class MemosCredentialsEffectHandler(
   private val loadCredentials: LoadCredentialsUseCase,
