+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/feature/memos/presentation/handler/MemosEffectHandler.kt
@@ -1,7 +1,9 @@
-package space.be1ski.vibits.shared.feature.memos.presentation
+package space.be1ski.vibits.shared.feature.memos.presentation.handler
 
 import kotlinx.coroutines.flow.Flow
 import space.be1ski.vibits.shared.core.elm.EffectHandler
+import space.be1ski.vibits.shared.feature.memos.presentation.MemosAction
+import space.be1ski.vibits.shared.feature.memos.presentation.MemosEffect
 
 class MemosEffectHandler(
   private val credentialsHandler: MemosCredentialsEffectHandler,
