+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/feature/mode/di/ModeSelectionFeatureFactory.kt
@@ -3,11 +3,11 @@ package space.be1ski.vibits.shared.feature.mode.di
 import space.be1ski.vibits.shared.core.elm.Feature
 import space.be1ski.vibits.shared.core.elm.FeatureImpl
 import space.be1ski.vibits.shared.feature.mode.presentation.ModeSelectionAction
-import space.be1ski.vibits.shared.feature.mode.presentation.ModeSelectionCredentialsEffectHandler
 import space.be1ski.vibits.shared.feature.mode.presentation.ModeSelectionEffect
-import space.be1ski.vibits.shared.feature.mode.presentation.ModeSelectionEffectHandler
-import space.be1ski.vibits.shared.feature.mode.presentation.ModeSelectionModeEffectHandler
 import space.be1ski.vibits.shared.feature.mode.presentation.ModeSelectionState
+import space.be1ski.vibits.shared.feature.mode.presentation.handler.ModeSelectionCredentialsEffectHandler
+import space.be1ski.vibits.shared.feature.mode.presentation.handler.ModeSelectionEffectHandler
+import space.be1ski.vibits.shared.feature.mode.presentation.handler.ModeSelectionModeEffectHandler
 import space.be1ski.vibits.shared.feature.mode.presentation.modeSelectionReducer
 
 fun createModeSelectionFeature(
