+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/feature/mode/presentation/handler/ModeSelectionCredentialsEffectHandler.kt
@@ -1,4 +1,4 @@
-package space.be1ski.vibits.shared.feature.mode.presentation
+package space.be1ski.vibits.shared.feature.mode.presentation.handler
 
 import kotlinx.coroutines.flow.Flow
 import space.be1ski.vibits.shared.core.elm.EffectHandler
@@ -9,6 +9,8 @@ import space.be1ski.vibits.shared.feature.auth.domain.usecase.InitializeCredenti
 import space.be1ski.vibits.shared.feature.auth.domain.usecase.LoadCredentialsUseCase
 import space.be1ski.vibits.shared.feature.auth.domain.usecase.SaveCredentialsUseCase
 import space.be1ski.vibits.shared.feature.memos.data.ConnectionTester
+import space.be1ski.vibits.shared.feature.mode.presentation.ModeSelectionAction
+import space.be1ski.vibits.shared.feature.mode.presentation.ModeSelectionEffect
 
 private const val TAG = "ModeCredentialsEffect"
 private const val LOG_URL_MAX_LENGTH = 20
