+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/feature/onboarding/di/OnboardingFeatureFactory.kt
@@ -3,12 +3,12 @@ package space.be1ski.vibits.shared.feature.onboarding.di
 import space.be1ski.vibits.shared.core.elm.Feature
 import space.be1ski.vibits.shared.core.elm.FeatureImpl
 import space.be1ski.vibits.shared.feature.onboarding.presentation.OnboardingAction
-import space.be1ski.vibits.shared.feature.onboarding.presentation.OnboardingCompletionEffectHandler
 import space.be1ski.vibits.shared.feature.onboarding.presentation.OnboardingEffect
-import space.be1ski.vibits.shared.feature.onboarding.presentation.OnboardingEffectHandler
-import space.be1ski.vibits.shared.feature.onboarding.presentation.OnboardingPresetsEffectHandler
-import space.be1ski.vibits.shared.feature.onboarding.presentation.OnboardingSetupEffectHandler
 import space.be1ski.vibits.shared.feature.onboarding.presentation.OnboardingState
+import space.be1ski.vibits.shared.feature.onboarding.presentation.handler.OnboardingCompletionEffectHandler
+import space.be1ski.vibits.shared.feature.onboarding.presentation.handler.OnboardingEffectHandler
+import space.be1ski.vibits.shared.feature.onboarding.presentation.handler.OnboardingPresetsEffectHandler
+import space.be1ski.vibits.shared.feature.onboarding.presentation.handler.OnboardingSetupEffectHandler
 import space.be1ski.vibits.shared.feature.onboarding.presentation.onboardingReducer
 
 fun createOnboardingFeature(
