+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/feature/onboarding/presentation/handler/OnboardingEffectHandler.kt
@@ -1,7 +1,9 @@
-package space.be1ski.vibits.shared.feature.onboarding.presentation
+package space.be1ski.vibits.shared.feature.onboarding.presentation.handler
 
 import kotlinx.coroutines.flow.Flow
 import space.be1ski.vibits.shared.core.elm.EffectHandler
+import space.be1ski.vibits.shared.feature.onboarding.presentation.OnboardingAction
+import space.be1ski.vibits.shared.feature.onboarding.presentation.OnboardingEffect
 
 class OnboardingEffectHandler(
   private val presetsHandler: OnboardingPresetsEffectHandler,
