+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/feature/onboarding/presentation/handler/OnboardingPresetsEffectHandler.kt
@@ -1,10 +1,12 @@
-package space.be1ski.vibits.shared.feature.onboarding.presentation
+package space.be1ski.vibits.shared.feature.onboarding.presentation.handler
 
 import kotlinx.coroutines.flow.Flow
 import space.be1ski.vibits.shared.core.elm.EffectHandler
 import space.be1ski.vibits.shared.core.elm.actions
 import space.be1ski.vibits.shared.core.logging.Log
 import space.be1ski.vibits.shared.feature.onboarding.domain.usecase.GetHabitPresetsUseCase
+import space.be1ski.vibits.shared.feature.onboarding.presentation.OnboardingAction
+import space.be1ski.vibits.shared.feature.onboarding.presentation.OnboardingEffect
 
 private const val TAG = "OnboardingPresetsEffect"
