+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/feature/onboarding/presentation/handler/OnboardingSetupEffectHandler.kt
@@ -1,4 +1,4 @@
-package space.be1ski.vibits.shared.feature.onboarding.presentation
+package space.be1ski.vibits.shared.feature.onboarding.presentation.handler
 
 import kotlinx.coroutines.flow.Flow
 import space.be1ski.vibits.shared.core.elm.EffectHandler
@@ -7,6 +7,8 @@ import space.be1ski.vibits.shared.core.logging.Log
 import space.be1ski.vibits.shared.core.platform.date.currentLocalDate
 import space.be1ski.vibits.shared.feature.onboarding.domain.usecase.CreateFirstCheckInUseCase
 import space.be1ski.vibits.shared.feature.onboarding.domain.usecase.CreateFirstHabitUseCase
+import space.be1ski.vibits.shared.feature.onboarding.presentation.OnboardingAction
+import space.be1ski.vibits.shared.feature.onboarding.presentation.OnboardingEffect
 
 private const val TAG = "OnboardingSetupEffect"
