+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/feature/settings/di/SettingsFeatureFactory.kt
@@ -5,12 +5,12 @@ import space.be1ski.vibits.shared.core.elm.Feature
 import space.be1ski.vibits.shared.core.elm.FeatureImpl
 import space.be1ski.vibits.shared.feature.mode.domain.model.AppMode
 import space.be1ski.vibits.shared.feature.settings.presentation.SettingsAction
-import space.be1ski.vibits.shared.feature.settings.presentation.SettingsCredentialsEffectHandler
 import space.be1ski.vibits.shared.feature.settings.presentation.SettingsEffect
-import space.be1ski.vibits.shared.feature.settings.presentation.SettingsEffectHandler
-import space.be1ski.vibits.shared.feature.settings.presentation.SettingsModeEffectHandler
-import space.be1ski.vibits.shared.feature.settings.presentation.SettingsPreferencesEffectHandler
 import space.be1ski.vibits.shared.feature.settings.presentation.SettingsState
+import space.be1ski.vibits.shared.feature.settings.presentation.handler.SettingsCredentialsEffectHandler
+import space.be1ski.vibits.shared.feature.settings.presentation.handler.SettingsEffectHandler
+import space.be1ski.vibits.shared.feature.settings.presentation.handler.SettingsModeEffectHandler
+import space.be1ski.vibits.shared.feature.settings.presentation.handler.SettingsPreferencesEffectHandler
 import space.be1ski.vibits.shared.feature.settings.presentation.settingsReducer
 
 fun createSettingsFeature(
