+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/feature/settings/presentation/handler/SettingsEffectHandler.kt
@@ -1,7 +1,9 @@
-package space.be1ski.vibits.shared.feature.settings.presentation
+package space.be1ski.vibits.shared.feature.settings.presentation.handler
 
 import kotlinx.coroutines.flow.Flow
 import space.be1ski.vibits.shared.core.elm.EffectHandler
+import space.be1ski.vibits.shared.feature.settings.presentation.SettingsAction
+import space.be1ski.vibits.shared.feature.settings.presentation.SettingsEffect
 
 class SettingsEffectHandler(
   private val credentialsHandler: SettingsCredentialsEffectHandler,
