+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/feature/settings/presentation/handler/SettingsPreferencesEffectHandler.kt
@@ -1,4 +1,4 @@
-package space.be1ski.vibits.shared.feature.settings.presentation
+package space.be1ski.vibits.shared.feature.settings.presentation.handler
 
 import kotlinx.coroutines.flow.Flow
 import space.be1ski.vibits.shared.core.elm.EffectHandler
@@ -6,6 +6,8 @@ import space.be1ski.vibits.shared.core.elm.sideEffect
 import space.be1ski.vibits.shared.core.logging.Log
 import space.be1ski.vibits.shared.feature.settings.domain.usecase.SaveLanguageUseCase
 import space.be1ski.vibits.shared.feature.settings.domain.usecase.SaveThemeUseCase
+import space.be1ski.vibits.shared.feature.settings.presentation.SettingsAction
+import space.be1ski.vibits.shared.feature.settings.presentation.SettingsEffect
 
 private const val TAG = "SettingsPreferencesEffect"
