+++ shared/src/commonTest/kotlin/space/be1ski/vibits/shared/feature/habits/presentation/HabitsEffectHandlerTest.kt
@@ -5,6 +5,10 @@ import kotlinx.coroutines.test.runTest
 import kotlinx.datetime.LocalDate
 import space.be1ski.vibits.shared.app.domain.model.ActivityMode
 import space.be1ski.vibits.shared.app.domain.model.ActivityRange
+import space.be1ski.vibits.shared.feature.habits.presentation.handler.HabitsActivityEffectHandler
+import space.be1ski.vibits.shared.feature.habits.presentation.handler.HabitsEffectHandler
+import space.be1ski.vibits.shared.feature.habits.presentation.handler.HabitsMemoEffectHandler
+import space.be1ski.vibits.shared.feature.habits.presentation.handler.HabitsRefreshEffectHandler
 import space.be1ski.vibits.shared.feature.memos.domain.model.Memo
 import space.be1ski.vibits.shared.feature.mode.domain.model.AppMode
 import space.be1ski.vibits.shared.test.FakeMemosRepository
