+++ shared/src/commonTest/kotlin/space/be1ski/vibits/shared/feature/memos/presentation/MemosEffectHandlerTest.kt
@@ -11,6 +11,10 @@ import space.be1ski.vibits.shared.feature.memos.domain.usecase.DeleteMemoUseCase
 import space.be1ski.vibits.shared.feature.memos.domain.usecase.LoadCachedMemosUseCase
 import space.be1ski.vibits.shared.feature.memos.domain.usecase.LoadMemosUseCase
 import space.be1ski.vibits.shared.feature.memos.domain.usecase.UpdateMemoUseCase
+import space.be1ski.vibits.shared.feature.memos.presentation.handler.MemosCredentialsEffectHandler
+import space.be1ski.vibits.shared.feature.memos.presentation.handler.MemosEffectHandler
+import space.be1ski.vibits.shared.feature.memos.presentation.handler.MemosLoadEffectHandler
+import space.be1ski.vibits.shared.feature.memos.presentation.handler.MemosWriteEffectHandler
 import space.be1ski.vibits.shared.test.FakeCredentialsRepository
 import space.be1ski.vibits.shared.test.FakeMemosRepository
 import kotlin.test.Test
